backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
requests==2.32.5
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
gunicorn==22.0.0